requirements.txt
gTTS==2.3.2
openai==0.27.8
PyAudio==0.2.13
pygame==2.5.1
python-dotenv==1.0.0
SpeechRecognition==3.10.0
